requirements.txt
absl-py==2.0.0
aiohttp==3.8.5
aioice==0.9.0
aiortc==1.5.0
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
av==10.0.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
boto3==1.28.59
botocore==1.31.59
cachetools==5.3.1
certifi==2022.12.7
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.7
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.4
cycler==0.10.0
dataclasses-json==0.6.1
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
dnspython==2.4.2
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.12.4
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.43.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
glob2==0.7
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-crc32c==1.5.0
greenlet==3.0.0
grpcio==1.59.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==2.10
ifaddr==0.2.0
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
labelImg==1.8.6
langchain==0.0.308
langsmith==0.0.42
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.26.0
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.16.0
openai==0.28.1
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
overrides==7.4.0
packaging==23.2
pafy==0.5.5
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.11.0
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
pyee==11.0.0
Pygments==2.16.1
pylibsrtp==0.8.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
roboflow==1.1.7
rpds-py==0.10.3
rsa==4.9
s3transfer==0.7.0
safetensors==0.3.3
sahi==0.11.14
scipy==1.11.3
seaborn==0.13.0
Send2Trash==1.8.2
shapely==2.0.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
stack-data==0.6.3
streamlit==1.27.2
streamlit-embedcode==0.1.2
streamlit-webrtc==0.47.1
supervision==0.14.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminado==0.17.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
tinycss2==1.2.1
tokenizers==0.14.0
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.1
transformers==4.34.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
ultralytics==8.0.192
uri-template==1.3.0
urllib3==1.26.17
validators==0.22.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==3.0.0
wget==3.2
widgetsnbextension==4.0.9
yarl==1.9.2
yolov5==7.0.12
zipp==3.17.0
plotly==5.18.0
